Remove wiki RPC subdomains and frontend WikiPage
